+++ examples/lib/stories/collision_detection/collision_detection.dart
@@ -4,6 +4,7 @@ import 'package:flame/game.dart';
 import '../../commons/commons.dart';
 import 'circles.dart';
 import 'multiple_shapes.dart';
+import 'only_shapes.dart';
 
 void addCollisionDetectionStories(Dashbook dashbook) {
   dashbook.storiesOf('Collision Detection')
@@ -16,5 +17,10 @@ void addCollisionDetectionStories(Dashbook dashbook) {
       'Multiple shapes',
       (_) => GameWidget(game: MultipleShapes()),
       codeLink: baseLink('collision_detection/multiple_shapes.dart'),
+    )
+    ..add(
+      'Shapes without components',
+      (_) => GameWidget(game: OnlyShapes()),
+      codeLink: baseLink('collision_detection/only_shapes.dart'),
     );
 }