+++ packages/flame/CHANGELOG.md
@@ -12,6 +12,8 @@
  - Abstracting the text api to allow custom text renderers on the framework
  - Set the same debug mode for children as for the parent when added
  - Fix camera projections when camera is zoomed
+ - Fix collision detection system with angle and parentAngle
+ - Fix rendering of shapes that aren't HitboxShape
 
 ## [1.0.0-rc9]
  - Fix input bug with other anchors than center